Add files via upload
The above the implementation of triangles were I am using the processing environment. What the program does is as we create points on the plane it will form a triangle by itself and calculates the area. Also, based on whether the triangle is in clockwise direction or counter clockwise direction it will color the triangle. For example, if the triangle formed by three points is clock wise than it may fill the triangle with green color and if it is counter clockwise then it may fill it with red color. Also, it calculates the internal angles of the triangle. It is an interesting project I would recommend to try yourself once. 

Important Note: It is one of my school project that but the work is completely and solely done my me.